Add files via upload
The "Installed Software Version Check" feature performs a crucial role in maintaining system health by identifying and reporting outdated software packages. It does so by executing a PowerShell command from within the Bash script, which checks for updates installed over three months ago using the Win32_QuickFixEngineering class. The results are clearly displayed, along with recommendations for action, thereby providing an easy-to-understand summary for system administrators and IT professionals.

This feature ensures that the system is up to date with the latest security patches and updates, which is important for protecting against vulnerabilities and ensuring optimal performance. By regularly running this check, administrators can proactively maintain the security and stability of their systems.